LocalStatusSQLite: Fix bug when deleting messages
The syntax was not right, and deleting messages from the LocalStatus
failed. (We passed in the full list of uids and we need to pass in one
uid at a time (as a tuple). Deleting messages works now.

+++ offlineimap/folder/LocalStatusSQLite.py
@@ -238,5 +238,5 @@ class LocalStatusSQLiteFolder(LocalStatusFolder):
             return
         for uid in uidlist:
             del(self.messagelist[uid])
-            self.sql_write('DELETE FROM status WHERE id=?',
-                             uidlist)
+            #TODO: we want a way to do executemany(.., uidlist) to delete all
+            self.sql_write('DELETE FROM status WHERE id=?', (uid, ))
